Move SettingsConfig
