+++ types/foundry/client/apps/index.d.ts
@@ -1,5 +1,4 @@
 import "./app.d.ts";
-import "./client-settings.d.ts";
 import "./compendium.d.ts";
 import "./form.d.ts";
 import "./forms/index.d.ts";
