+++ types/foundry/client/apps/sidebar/apps/index.d.ts
@@ -1,2 +1,3 @@
 import "./chat-popout.d.ts";
+import "./client-settings.d.ts";
 import "./keybindings-config.d.ts";
